Add home view for avatar and bison Protect routes by role Logic for login by role Data for users and packages Add phone field for user interface
