+++ src/views/HomeBisonView.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+<h1>Home for Bison</h1>
+</template>
+
+<style scoped>
+
+</style>
